+++ src/environments/environment.prod.ts
@@ -1,6 +1,5 @@
 export const environment = {
   production: true,
-  CSP_NONCE: 'cHJhaGEgYnJubyBvc3RyYXZhIHBhcmR1YmljZSBtb3N0IHBsemVuIGppaGxhdmE=',
   SERVER_ORIGIN: 'https://recipeappexpress-7211.rostiapp.cz',
   LOGIN_PAGE_URL: 'https://recipeappexpress-7211.rostiapp.cz/login',
   LOGOUT_PAGE_URL: 'https://recipeappexpress-7211.rostiapp.cz/logout',