+++ src/environments/environment.ts
@@ -1,6 +1,5 @@
 export const environment = {
   production: false,
-  CSP_NONCE: 'cHJhaGEgYnJubyBvc3RyYXZhIHBhcmR1YmljZSBtb3N0IHBsemVuIGppaGxhdmE=',
   SERVER_ORIGIN: 'http://localhost:3000',
   LOGIN_PAGE_URL: 'http://localhost:3000/login',
   LOGOUT_PAGE_URL: 'http://localhost:3000/logout',