+++ src/main.ts
@@ -1,4 +1,4 @@
-import { CSP_NONCE, enableProdMode } from '@angular/core';
+import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
@@ -9,12 +9,5 @@ if (environment.production) {
 }
 
 platformBrowserDynamic()
-  .bootstrapModule(AppModule, {
-    providers: [
-      {
-        provide: CSP_NONCE,
-        useValue: environment.CSP_NONCE,
-      },
-    ],
-  })
+  .bootstrapModule(AppModule)
   .catch((err) => console.error(err));